// web/src/app/api/model/section-response.ts
/**
 * Simplicity training API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SectionResponse { 
    body?: string;
    position?: number;
    sectionId?: string;
    title?: string;
}